+++ lib/Service/DataloadService.php
@@ -232,7 +232,9 @@ class DataloadService
                 $error = $error + 1;
                 continue;
             }
-            // if data source only delivers 2 columns, the value needs to be in the last one
+			$this->logger->info('before: ' .$row[0]. '_' . $row[1]. '_' . $row[2]);
+			// if data source only delivers 2 columns, the value is moved to the last one
+			// resulting in xxx, null , 123
             if (count($row) === 2) {
                 $row[2] = $row[1];
                 $row[1] = null;